LayoutTests/platform/blackberry/fast/dom/elementFromPoint-relative-to-viewport-expected.txt

This test document.elementFromPoint is evaluated in with respect to the viewport, not the document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS unscrolledBoxInitial is '0'
PASS scrolledDownBoxInitial is '5'
FAIL scrolledRightBoxInitial should be 3. Was 0.
FAIL scrolledDownAndRightBoxInitial should be 8. Was 5.
PASS unscrolledBoxOffscreen is '0'
PASS scrolledDownBoxOffscreen is '5'
FAIL scrolledRightBoxOffscreen should be 3. Was 0.
FAIL scrolledDownAndRightBoxOffscreen should be 8. Was 5.
PASS unscrolledBoxInitial is '0'
PASS scrolledDownBoxInitial is '5'
FAIL scrolledRightBoxInitial should be 3. Was 0.
FAIL scrolledDownAndRightBoxInitial should be 8. Was 5.
PASS successfullyParsed is true

TEST COMPLETE
